Rename bootstrap package in common
