+++ tests/http_app/conftest.py
@@ -10,6 +10,6 @@ from http_app import create_app
 @pytest.fixture(scope="function")
 def testapp(test_config) -> Iterator[FastAPI]:
     # We don't need the storage to test the HTTP app
-    with patch("bootstrap.bootstrap.init_storage", return_value=None):
+    with patch("common.bootstrap.init_storage", return_value=None):
         app = create_app(test_config=test_config)
         yield app
