+++ tests/storage/conftest.py
@@ -4,7 +4,7 @@ import pytest
 from sqlalchemy.orm import clear_mappers
 from sqlalchemy_bind_manager import SQLAlchemyBindManager, SQLAlchemyConfig
 
-from bootstrap.storage.SQLAlchemy import init_tables
+from common.storage.SQLAlchemy import init_tables
 
 
 @pytest.fixture(scope="function")
